chore: only subscription to myUpdates once
also may fix #1062
