+++ app/graphql/ln-update-context.ts
@@ -1,0 +1,7 @@
+import { createContext, useContext } from "react"
+
+const LnUpdateHash = createContext<string>("")
+
+export const LnUpdateHashProvider = LnUpdateHash.Provider
+
+export const useLnUpdateHash = () => useContext(LnUpdateHash)
